Added obj_architect.object
This code is for demonstration purposes only.

This 'object', which functions like a class in traditional languages, is
what procedurally generates the dungeon in the mini-game Treasure Hunt.

A dungeon (or 1 full level), is comprised of 3 zones, which connect to
each other through the red and green tiles, and have some basic pickups,
which as of yet do nothing.