Added 2 scripts (scr_test_overlap, scr_create_room)
This code is for demonstration purposes only.

The first script, tests the provided room, against all placed rooms in
its zone. If it overlaps with any of them, it is discarded. If not, it
is added to the list, and then created by the script scr_create_room.

The second script creates the room according to the dimension provided
by the object being passed (room 'r'). It creates wall tiles on the
outer-most tiles of the room, and creates floor tiles for the rest; it
also clears these tiles in the mp_grid's corresponding cell, which
allows the player to move around on floor tiles.